+++ src/test/java/com/soter/idomok/IdomokTest.java
@@ -31,17 +31,17 @@ public class IdomokTest {
 
     @Test
     void testGombTerfogat(){
-        idom = new Gomb((int) 1.0);
+        idom = new Gomb(1.0);
         assertEquals(4.188, idom.terfogat(), 0.001);
     }
 //
     @Test
     void testGombSuly(){
-        idom = new Gomb((int) 1.0);
+        idom = new Gomb(1.0);
         double d = idom.suly();
         assertEquals(3.141, idom.suly(), 0.001);
     }
-//
+
 //    @Test
 //    void testHasabTerfogat(){
 //        idom = new Hasab(1.0, 1.0, 1.0);
